refactor: rename and move scenario API interfaces to new package
Rename ConditionalScenarioConditionStep to ConditionBehaviorChain,
ConditionalScenarioActionStep to ConditionBehaviorChainActionStep,
and move all scenario interfaces (ConditionBehaviorChain,
ConditionBehaviorChainActionStep, ScenarioCondition) into the new
io.camunda.process.test.api.scenario package.
