+++ testing/camunda-process-test-java/src/main/java/io/camunda/process/test/api/scenario/ScenarioCondition.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.process.test.api;
+package io.camunda.process.test.api.scenario;
 
 /**
  * Functional interface representing a scenario condition that is evaluated periodically by the
